+++ src/ProjectNeon/Assets/Features/Combatants/IStats.cs
@@ -1,10 +1,11 @@
 ﻿
+// @todo #1:60min Generify this using StatType
 public interface IStats
 {
-    int MaxHP { get; }
-    int MaxShield { get; }
-    int Attack { get; }
-    int Magic { get; }
+    float MaxHP { get; }
+    float MaxShield { get; }
+    float Attack { get; }
+    float Magic { get; }
     float Armor { get; }
     float Resistance { get; }
     IResourceType[] ResourceTypes { get; }
